+++ package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@lottiefiles/react-lottie-player": "^3.5.4",
+    "@uidotdev/usehooks": "^2.4.1",
     "axios": "^1.7.2",
     "next": "14.2.4",
     "openai": "^4.50.0",
