+++ src/app/gpt/route.ts
@@ -5,7 +5,7 @@ import { Validators } from "../utils/validation";
 
 export const dynamic = "force-dynamic"; // defaults to auto
 
-export async function POST(request: Request): Promise<GptResponse | unknown> {
+export async function POST(request: Request): Promise<GptResponse | any> {
   const body: GptRequestBody = await request.json();
   const errors = Validators.GptBodyValidator(body as GptRequestBody);
